pyproject.toml
[project]
name = "o1-replication-journey-tree-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.39.0",
    "streamlit-flow-component>=1.5.0",
    "pydantic>=2.9.2",
    "typer>=0.12.5",
]
